Append UUID to resource name to guarantee that it is unique
Append a UUID to the Jabber resource name so that it is non-empty and
unique. Since Jabber uses resource names to identify distinct sessions
from the same user, this will make it so that connecting to an existing
resource name will not boot other clients off.

See <https://sympa.ligo.org/wws/arc/daswg/2016-05/msg00209.html>.

Signed-off-by: Alexander Pace <alexander.pace@client-104-39-69-143.mobility.psu.edu>